feat: Update API call in Gift screen to fetch products by type
This commit updates the API call in the Gift screen to fetch products based on their type. The previous code was fetching products of types 'higiene', 'Roupas', 'Alimentos', and 'pet', but it has been updated to fetch products of types 'hygiene' and 'pet' instead. This change improves the accuracy of the fetched data.