refactor: Add GiftCart component and context
Add the GiftCart component and GiftContext to the project. The GiftCart component displays a gift icon and the number of gifts in the cart. The GiftContext provides functionality to add and remove gifts from the cart. This enhancement improves the user experience by providing visual feedback on the number of gifts in the cart.